+++ demo/src/app/api-client-module/auth.model.ts
@@ -1,0 +1,7 @@
+import { Auth } from 'soft-ngx';
+
+export interface Auth extends Auth {
+  is_admin: boolean;
+  user_id: number;
+  display_name: string;
+}