+++ src/api-client/api-client.module.ts
@@ -8,7 +8,6 @@ import { windowToken, getWindow } from '../window';
 
 import { ApiClientConfig } from './api-client.config';
 import { ApiClientService } from './api-client.service';
-import { AuthHelperService } from './auth-helper.service';
 import { AuthService } from './auth.service';
 import { StorageService } from './storage.service';
 import { userConfigToken } from './user-config.token';
@@ -29,7 +28,6 @@ export class ApiClientModule {
         { provide: userConfigToken, useValue: config },
         { provide: windowToken, useFactory: getWindow },
         ApiClientService,
-        AuthHelperService,
         AuthService,
         StorageService
       ]