+++ src/api-client/auth.model.ts
@@ -1,4 +1,4 @@
-export class Auth {
+export interface Auth {
   access_token: string;
   token_type: string;
   expires_in: number;