+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 ## Changelog
 
-### 3 Jan, 2018
+### v0.2.0
+* Remove AuthHelperService
+* Set additional auth response data via ApiClientConfig
+* Export ModalComponent
+
+### v0.1.0
 * Create lib based on robisim74/angular-library-starter (Nov 2, 2017)